Queries Redesign
Moved functions around, renamed:
get_mediaitem_by_mediaitem_id to get_general_item_info
get_mediaitem_by_barcode to get_copy_info
Added new files and folders, moved some of Queries' functions into them.
It is tested (a bit) and they seem to be working.

Partially alphebetized the functions in the three new files, with the
basic function on top, with a comment block.

Added comments to testfile and testdata for functions not yet tested,
and reordered the function's testing to match the ordering in the files.

Added some not-so-helpful comments to testfile.